+++ src/command_modules/azure-cli-container/azure/cli/command_modules/container/_params.py
@@ -9,7 +9,8 @@ from azure.cli.core.commands.parameters import (get_enum_type,
                                                 get_location_type,
                                                 resource_group_name_type)
 from azure.cli.core.commands.validators import get_default_location_from_resource_group
-from azure.mgmt.containerinstance.models import (ContainerGroupRestartPolicy, OperatingSystemTypes)
+from azure.mgmt.containerinstance.models import (
+    ContainerGroupRestartPolicy, OperatingSystemTypes, ContainerNetworkProtocol)
 from ._validators import validate_volume_mount_path, validate_secrets, validate_gitrepo_directory
 
 # pylint: disable=line-too-long
@@ -49,6 +50,7 @@ def load_arguments(self, _):
         c.argument('os_type', arg_type=get_enum_type(OperatingSystemTypes), help='The OS type of the containers')
         c.argument('ip_address', arg_type=get_enum_type(IP_ADDRESS_TYPES), help='The IP address type of the container group')
         c.argument('ports', type=int, nargs='+', default=[80], help='The ports to open')
+        c.argument('protocol', arg_type=get_enum_type(ContainerNetworkProtocol), help='The network protocol to use')
         c.argument('dns_name_label', help='The dns name label for container group with public IP')
         c.argument('restart_policy', arg_type=get_enum_type(ContainerGroupRestartPolicy), help='Restart policy for all containers within the container group')
         c.argument('command_line', help='The command line to run when the container is started, e.g. \'/bin/bash -c myscript.sh\'')
@@ -69,8 +71,7 @@ def load_arguments(self, _):
         c.argument('azure_file_volume_mount_path', validator=validate_volume_mount_path, help="The path within the container where the azure file volume should be mounted. Must not contain colon ':'.")
 
     with self.argument_context('container create', arg_group='Log Analytics') as c:
-        c.argument('log_analytics_workspace', help='The Log Analytics workspace. Must be in the currently set subscription.')
-        c.argument('log_analytics_workspace_id', help='The Log Analytics workspace id.')
+        c.argument('log_analytics_workspace', help='The Log Analytics workspace name or id. If a name is specified, it must be within the currently set subscription.')
         c.argument('log_analytics_workspace_key', help='The Log Analytics workspace key.')
 
     with self.argument_context('container create', arg_group='Git Repo Volume') as c: