+++ src/command_modules/azure-cli-container/HISTORY.rst
@@ -6,7 +6,8 @@ Release History
 0.2.0
 +++++
 * Default to long running operation for `az container create`
-* Add Log Analytics parameters '--log-analytics-workspace' '--log-analytics-workspace-id' and '--log-analytics-workspace-key'
+* Add Log Analytics parameters '--log-analytics-workspace' and '--log-analytics-workspace-key'
+* Add --protocol parameter to specify which network protocol to use
 
 0.1.24
 ++++++